Fix: move early env processing into utils to avoid side effects.
